refactor: Rename `…gins/aws/package/lib/mergeCustomProviderResources.js`
