+++ lib/plugins/aws/package/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const path = require('path');
-const mergeCustomProviderResources = require('./lib/mergeCustomProviderResources');
+const mergeCustomProviderResources = require('./lib/merge-custom-provider-resources');
 const stripNullPropsFromTemplateResources = require('./lib/stripNullPropsFromTemplateResources');
 const generateArtifactDirectoryName = require('./lib/generate-artifact-directory-name');
 const generateCoreTemplate = require('./lib/generate-core-template');
