+++ lib/plugins/aws/package/lib/merge-custom-provider-resources.js
