+++ test/unit/lib/plugins/aws/package/lib/merge-custom-provider-resources.test.js
